feat: adjustments made on the customers list screen
