+++ frontend/src/routes/routes.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter, Navigate } from 'react-router-dom'
 import { RootLayout } from '../layouts/RootLayout/RootLayout'
 import { PetsListing } from '../pages/Pets/PetsListing'
 import { PetDetail } from '../pages/Pets/PetDetail'
-import { CustomersListing } from '../pages/Customers/CustomersListing'
+import { CustomersListing } from '../pages/Customers/CustomersListing/CustomersListing'
 import { CustomerDetail } from '../pages/Customers/CustomerDetail/CustomerDetail'
 
 export const routes = createBrowserRouter([
